requirements.txt
-i https://pypi.org/simple
aspose-cells==20.2.0
jpype1==1.1.1
